+++ CoEvolution Project/Texturesheet.cpp
@@ -3,8 +3,8 @@
 Texturesheet::Texturesheet(sf::Image texturesheet, sf::Vector2u texturesize) : _texturesize(texturesize) 
 {
 	this->_sheetsize = sf::Vector2u(texturesheet.getSize().x / _texturesize.x, texturesheet.getSize().y / _texturesize.y);
-	for (int y = 0; y < _sheetsize.y;y++) {
-		for (int x = 0; x < _sheetsize.x; x++) {
+	for (int y = 0; y < (int)_sheetsize.y;y++) {
+		for (int x = 0; x < (int)_sheetsize.x; x++) {
 			this->_textures.push_back(sf::Texture());
 			this->_textures.back().loadFromImage(texturesheet, sf::IntRect(x*_texturesize.x, y*_texturesize.y, _texturesize.x, _texturesize.y));
 		}
